Workaround to fix semicolon syntax issue with keyframes
Context: Lessc is adding a semicolon at the end of keyframes definition, which is
invalid css and causes issues on some devices.  This also creates issues
if you use anything that parses your css.

This adds a noop rule at the end of each keyframes definition, leading
to valid css